Re-added Herald service UUID to advert data
We were relying on GATT but Herald on iOS and Android (correctly) filters devices without a Herald UUID in the advert. I've had to remove TxPower from the advert on zephyr as a result. I've logged an issue to re add it once we have a short service uuid.
See https://github.com/vmware/herald-for-cpp/issues/27
Fixes #26
Signed-off-by: Adam Fowler <adamfowleruk@gmail.com>